[CI] Use git merge instead of rebase in sync-upstream workflow
**Context**
Workflow create-feature-branch wasn't get triggered when v0.4.8 is pushed.

**RCA**
GitHub Actions will only trigger on a pushed tag if the tagged commit contains the workflow file itself.

For instance, because we have sync upstream which rebase all our own workflows on top of the upstream commits. A tag such as v0.4.8 is associated with an older commit, the workflow won't get triggered.

**Changes**
- Use git merge instead of git rebase in sync-upstream.yml